requirements.txt
langchain
langchain-core
streamlit
boto3
langchain_community
python-dotenv